+++ lib/kfx_dependency_injection.dart
@@ -5,6 +5,7 @@ import 'kfx_dependency_injection/platform_info.dart';
 part 'kfx_dependency_injection/i_must_be.dart';
 part 'kfx_dependency_injection/i_platform_info.dart';
 part 'kfx_dependency_injection/invalid_registration_modal_for_type_exception.dart';
+part 'kfx_dependency_injection/platform_design_system.dart';
 part 'kfx_dependency_injection/platform_host.dart';
 part 'kfx_dependency_injection/platform_media.dart';
 part 'kfx_dependency_injection/service_already_registered_exception.dart';