+++ lib/kfx_dependency_injection/platform_design_system.dart
@@ -1,0 +1,16 @@
+part of kfx_dependency_injection;
+
+/// The design system used by the platform host
+enum PlatformDesignSystem {
+  /// Android or Linux
+  materialDesign,
+
+  /// MacOS, iPhone or iPad
+  appleHumanIntercace,
+
+  /// Windows machine
+  fluentDesign,
+
+  /// Unable to determine the host or design system
+  unknown,
+}